Add files via upload
Npte that for this script to work, you need to have turtlesim running in advance

NOTE2: This code does not fully work. After hours of testing and finetuning, i still am not able to get the gravity to work after halfway. I would love to know what exactly I was missing with this script :)